// src/data/eu-notice-questions/types.ts
export type {
  Question,
  QuestionOption,
  ShowIfCondition,
  FlagCondition,
  QuestionInputType,
} from "@/data/ropa-questions/types";

export type EuFrameworkCode =
  | "EU_GDPR"
  | "UK_GDPR"
  | "CH_FADP"
  | "BR_LGPD"
  | "JP_APPI"
  | "IN_DPDPA"
  | "ZA_POPIA"
  | "CA_PIPEDA"
  | "AU_PRIVACY"
  | "KR_PIPA"
  | "SG_PDPA"
  | "AE_PDPL";